Vehicle : Ajout d'ExceptionHandler et statuts HTTP
Modification du VehicleService pour intégrer des Exceptions, et traitement de ces exception avec ExceptionHandlers dans le controller. Gestion des statuts retournés par le controller.

+++ src/main/java/com/ftgcar/services/VehicleService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.ftgcar.dao.VehicleRepository;
 import com.ftgcar.dto.VehicleDto;
 import com.ftgcar.entity.Vehicle;
+import com.ftgcar.exception.AlreadyExistsException;
 import com.ftgcar.exception.NotFoundException;
 import com.ftgcar.mapper.VehicleMapper;
 
@@ -24,9 +25,22 @@ public class VehicleService {
         this.vehicleRepository = vehicleRepository;
     }
 
-    public VehicleDto addVehicle(VehicleDto vehicleDto) {
+    public VehicleDto addVehicle(VehicleDto vehicleDto) throws AlreadyExistsException {
+        checkIfVehicleAlreadyExistsForAdd(vehicleDto);
         vehicleRepository.save(vehicleMapper.vehicleDtoToVehicle(vehicleDto));
         return vehicleDto;
+    }
+
+    private void checkIfVehicleAlreadyExistsForAdd(VehicleDto vehicleDto) throws AlreadyExistsException {
+        Optional<Vehicle> existingVehicleById = vehicleRepository.findById(vehicleDto.id());
+        if (existingVehicleById.isPresent()) {
+            throw new AlreadyExistsException(String.format("Le véhicule à l'id %s existe déjà.", vehicleDto.id()));
+        }
+        Optional<Vehicle> existingVehicleByNumberplate = vehicleRepository.findByNumberplate(vehicleDto.numberplate());
+        if (existingVehicleByNumberplate.isPresent()) {
+            throw new AlreadyExistsException(
+                    String.format("Le véhicule avec l'immatriculation %s existe déjà.", vehicleDto.numberplate()));
+        }
     }
 
     public List<VehicleDto> findAllVehicles() {
@@ -39,21 +53,30 @@ public class VehicleService {
     public VehicleDto findVehicleById(long id) throws NotFoundException {
         Optional<Vehicle> existingVehicle = vehicleRepository.findById(id);
         if (existingVehicle.isEmpty()) {
-            throw new NotFoundException("Sauf erreur de ma part, le véhicule demandé n'existe pas.");
+            throw new NotFoundException("Le véhicule demandé n'existe pas.");
         }
         return vehicleMapper.vehicleToVehicleDto(existingVehicle.get());
     }
 
     public void deleteVehicleById(Long id) {
-        vehicleRepository.deleteById(id);
+        Optional<Vehicle> existingVehicle = vehicleRepository.findById(id);
+        if (existingVehicle.isPresent()) {
+            vehicleRepository.deleteById(id);
+        }
     }
 
-    public VehicleDto updateVehicleById(Long id, VehicleDto vehicleDto) {
-        Optional<Vehicle> vehicleToUpdate = vehicleRepository.findById(id);
+    public VehicleDto updateVehicle(VehicleDto vehicleDto) throws AlreadyExistsException {
+        Optional<Vehicle> vehicleToUpdate = vehicleRepository.findById(vehicleDto.id());
         if (vehicleToUpdate.isPresent()) {
+            Optional<Vehicle> vehicleExistingWithNumberplate = vehicleRepository
+                    .findByNumberplate(vehicleDto.numberplate());
+            if (vehicleExistingWithNumberplate.isPresent()
+                    && !vehicleExistingWithNumberplate.get().getId().equals(vehicleDto.id())) {
+                throw new AlreadyExistsException(
+                        String.format("L'immatriculation %s existe déjà.", vehicleDto.numberplate()));
+            }
             return vehicleMapper.vehicleToVehicleDto(vehicleMapper.updateVehicle(vehicleDto, vehicleToUpdate.get()));
         }
         return addVehicle(vehicleDto);
     }
-
 }
